requirements.txt
django-debug-toolbar==3.8.1
django-extensions==3.2.1
django-phac_aspc-helpers==0.4.3
django-versionator==0.0.1b0
Django==4.1.9
Faker==15.1.1
gunicorn==20.1.0
ipython==8.10.0
Jinja2==3.1.2
psycopg2-binary==2.9.5
python-decouple==3.6
pytz==2022.2.1
rules==3.3
whitenoise==6.4.0